Reduce code duplication further.
Given that onPause() method needs to also disable DM mode,
I moved the logic around some more so it can be reused
there and by the DirectManipulationHandler class.

Also fixed a bug I found that didn't consume nudge and
rotation events when no handler delegate was provided for
them. Not consuming the nudge events means that in DM mode
user can nudge out of the view that should be directly
manipulated. The rotation one doesn't cause an issue
as far as I can tell but I am consuming them because
that is what the javadoc claims to do (provide a "noop"
impl if a delegate not provided = consume event and
do nothing).

BUG: 156978213
Test: make, install, run code
Change-Id: I417b2dcbd5e0fd36e11db58681a8baff9bbfcf8e