Add direct manipulation sys window widgets.
In particular, adding 2 SeekBars and 2 RadialTimePickerViews.
2 because in a system window, a given View can be marked as
"supports direct manipulation mode" or not. So we add 1 of
each thing that supports DM mode and one that doesn't.

For the ones that support, entering DM mode allows rotary
controls to be used to manipulate the value.

For the ones that don't, the event that would have entered
DM mode instead triggers an ACTION_CLICK which in our
implementation changes the background color on every other invocation.

Test: manual, build, install, run the app
BUG: 153888753
Change-Id: If1d4de9def33c8bf062e6f0ce202b99797e2d014