Make scroll fragment more challenging.
Add a gap between focusable elements to test the transition from an
element being focused to the scrollable container being focused and
vice versa.

Add non-focusable content at the top and bottom of the list to make
sure you can scroll to the top and bottom in this scenario.

Add other focusable views in the same focus area with the scrollable
view to test rotation into and out of scrollable containers.

Remove content description indicating that the CarUiRecyclerView is
scrollable. CarUiRecyclerViews are now opt-out rather than opt-in.

Bug: 154540419
Test: manual
Change-Id: Iea3723dedd255817e4c6fc5c6b0487bccbe2604c