Add a WebView to the reference app
Add another fragment to the RotaryPlayground reference app for testing
rotary navigation into, out of, and within WebViews.

The WebView's content is chosen to include sections without any
focusable elements.

Includes a button to toggle the visibility of two buttons in the same
focus area with the WebView. This enables testing:
- nudging into a focus area with only a WebView without any visible
  focusable elements
- rotating into a WebView without any visible focusable elements.

Bug: 159936127
Test: manual
Change-Id: I1273e32eec49ad0983f8875c4ad95938c7e50211