Add generate-overlays.py
Similar to generate-overlayable.py, but for the <overlay> entry.
This assumes that the resources that the RRO uses to replace
app resources are named exactly the same as the app's resources.

The user can also supply the app's res folder to the script, in
which case it will know to ignore any resources that are not also
present in the app.

Fixes: 169707403
Test: Manually
Change-Id: I33340f8f711fefa0166c7d775881eab82d1f8907